pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-validator"
version = "0.2.0"
description = "Rule-based Markdown document validation for static site content."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Matt Briggs" }]
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.3.7",
    "lxml>=5.3.0",
    "nltk>=3.7",
    "html2text>=2024.2.26",
    "PyYAML>=6.0.2",
    "pydantic>=2.0",
    "click>=8.1.3",
    "colorama>=0.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[project.scripts]
md-validate = "markdown_validator.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=markdown_validator --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["markdown_validator"]
omit = [
    "markdown_validator/cli/repl.py",
    "markdown_validator/cli/main.py",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
